requirements.txt
# Django Core
Django==2.2.28
python-dotenv==1.1.1

# Database & ORM
djongo==1.2.31
pymongo==4.15.2
dnspython==2.8.0

# AI & ML
openai==1.12.0
google-generativeai==0.3.2
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20

# Vector Database
chromadb==0.4.22
sentence-transformers==2.3.1

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0

# OCR & Document Processing
pytesseract==0.3.10
python-magic==0.4.27

# Async Task Processing
celery==5.3.4
redis==5.0.1
kombu==5.3.5

# API & Web
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2025.2
typing-extensions==4.15.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1

# Security
cryptography==42.0.2

# Monitoring & Logging
sentry-sdk==1.40.0
